Keep browserstack code co-located
"Faraday" helper was included at a global level, but is only needed
inside a single module. Lets inline it there.
