+++ testing/features/support/helpers/all.rb
@@ -2,7 +2,6 @@
 
 require_relative "env_variables"
 require_relative "drivers"
-require_relative "faraday"
 require_relative "javascript"
 require_relative "methods"
 require_relative "page"
